requirements.txt
python-dotenv==0.20.0
python-telegram-bot==13.13
redis==4.3.4
